Version 4.2 (#111)
* upgrade node version in workflow

* packge upgrade

* update eslint

* Feat defaults (#106)

* add fallback for tenant default (unmanaged room)

* fix file size conversion

* update turn hashing

* add reactions

* yarn upgrade

* jws update

* Use tenant default for managedRoom if empty role is set

* Fix default room settings (#107)

* Fix default room settings

* Mediasoup 3.19.14

* Fix fallback for maxFileSize

* Move to node 22

* Move to node 24

* Fix calc of file size and tracker

* Fix calc of file size and tracker v2

* Fix calc of file size and tracker v3

* Mediasoup 3.19.14 - yarn.lock

---------


* Fix webtorrent - use turn v2

* Fix webtorrent - use turn v2 (#108)


* Upgrade socket.io socket.io-client edumeet-common

* Upgrade debug. Fix logger because of pino

* Update yarn.lock

* Astagor fix 4.2 rc6 (#109)

* Fix webtorrent - use turn v2

* Upgrade socket.io socket.io-client edumeet-common

* Upgrade debug. Fix logger because of pino

* Update yarn.lock

---------


* Fix interactive server unlink

* Update edumeet commons

* Fix socket connection to management server

* Fix socket connection to management server v2

* Fix socket connection to management server v3

* Fix socket connection to management server v4

* Fix socket connection to management server v5

* Fix dobule reAuth to management server

* Fix lint

* Fix reAuth in constructor

* Add debug logs to reAuth

* Fix use authenticateLocal to get fresh token every hour

---------

+++ src/interactiveServer.ts
@@ -134,8 +134,12 @@ export const interactiveServer = (serverManager: ServerManager, managementServic
 	});
 
 	try {
-		fs.unlinkSync(SOCKET_PATH);
-	} catch (error) {}
+		if (fs.existsSync(SOCKET_PATH)) {
+			fs.unlinkSync(SOCKET_PATH);
+		}
+	} catch (error) {
+		logger.debug({ err: error }, 'InteractiveServer listening [error: %o}');
+	}
 
 	server.listen(SOCKET_PATH, () => {
 		logger.debug('InteractiveServer listening [socket: %s}', SOCKET_PATH);
